Breaking! Make coords a string, improve exec perf ~4x
It is much faster because String coords will be added to the constant
pool of the class, and the bytecode will just reference it, instead of
having bytecode continuously creating arrays.
Since they are string on the constant pool they will be automatically
interned when loading the classes, so no need for a coord cache neither.